requirements.txt
south
uWSGI
python-openid
django-bootstrap-form
django-crispy-forms
django-avatar
django-filter
pil
psycopg2
-e ext/django-allauth
djangorestframework
markdown
git+git://github.com/SiteSupport/gevent.git@1.0rc2#egg=gevent
git+git://github.com/django/django.git@stable/1.5.x#egg=django
